requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
colorama==0.3.9
configparser==3.5.0
Django==1.9
django-tastypie==0.14.2
docutils==0.14
enum34==1.1.6
futures==3.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.3
python-dateutil==2.7.3
python-mimeparse==1.6.0
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
